adding argparser to config manager
