+++ app_configs/config_manager.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import json
 import os
 import pathlib
+import argparse
 
 
 class ConfigManager:
@@ -30,10 +31,9 @@ class ConfigManager:
             ConfigManager.__instance = self
 
         self.config_dict = {}
-        self.arg_dict = {}
-        self.load_config_file(f'{str(pathlib.Path(__file__).parent.absolute())}/config.json')
+        self._load_config_file(f'{str(pathlib.Path(__file__).parent.absolute())}/config.json')
 
-    def load_config_file(self, filepath: str) -> None:
+    def _load_config_file(self, filepath: str) -> None:
         """
         Attempts to load config file located at filepath.
         """
@@ -48,30 +48,45 @@ class ConfigManager:
         """
         @return: The configured database name or default url
         """
-        return self.config_dict['db_url'] if 'db_url' in self.config_dict else 'sqlite:///bodega_discord.db'
+        return self.config_dict['db_url'] or 'sqlite:///bodega_discord.db'
 
 
     def get_bot_token(self) -> str:
         """
         @return: The configured bot token or default test bodega token
         """
-        bot_token = ""
-        if 'bot_token' in self.config_dict:
-            if self.config_dict['bot_token']:
-                bot_token = self.config_dict['bot_token']
-
-        if 'bot_token' in self.arg_dict:
-            if self.arg_dict['bot_token']:
-                bot_token = self.arg_dict['bot_token']
-
-        return bot_token
+        return self.config_dict['bot_token'] or ""
 
     def inject_parsed_arguments(self, arguments: dict) -> None:
         """
         Populates private argument dictionary.
         """
+
+        self._load_config_file(arguments["config_path"])
         for argument, value in arguments.items():
-            self.arg_dict[argument] = value
+            if argument not in self.config_dict:
+                self.config_dict[argument] = value
+
+def user_input() -> None:
+    """
+    Collects all the user inputs from the CLI
+    """
+    parser = argparse.ArgumentParser(
+        description="smart-nine generates an instagram user's smart top 9 photograph collage",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        fromfile_prefix_chars='@')
+
+    parser.add_argument('-bot_token', '-b', type=str, default="", help='Discord Bot token')
+    parser.add_argument('--db_url', '-d', type=str, default="", help='Database url or connection string.')
+    parser.add_argument('--config-path', '--config_path', '-c', type=str, default="", help='Filepath to configuration file.')
+    args = parser.parse_args()
+    ConfigManager.get_instance().inject_parsed_arguments(args.__dict__)
+
+    if ConfigManager.get_instance().get_bot_token() == "":
+        parser.print_help()
+        raise ValueError('Must provide a Discord Bot token OR a config file containing a Bot token')
+    
+
